马蜂窝/requirements.txt
PyExecJS==1.5.1
requests==2.25.1